Add the kotlin-api-client as a submodule to the project (#1)
This adds the kotlin-api-client to the POC so that we should be able to make API calls. The API client needed to be generated based on the dependency updates of the openapi-client-generator which didn't make it to an official release yet. Therefore it was not added as a Git submodule. More information about the changes on the generator project here: https://github.com/OpenAPITools/openapi-generator/pull/9232